factor out the reusable bits
